Store Limits - Part 1

+++ api/store.js
@@ -5,11 +5,15 @@ const fs = require("fs");
 const ejs = require("ejs");
 
 module.exports.load = async function(app, db) {
+  let maxram = 1024;
+  let maxcpu = 200;
+  let maxservers = 4;
+  let maxdisk = 10000;
   app.get("/buyram", async (req, res) => {
     let newsettings = await enabledCheck(req, res);
     if (newsettings) {
       let amount = req.query.amount;
-
+      
       if (!amount) return res.send("missing amount");
 
       amount = parseFloat(amount);
@@ -36,7 +40,7 @@ module.exports.load = async function(app, db) {
 
       let newusercoins = usercoins - cost;
       let newram = ramcap + amount;
-
+      if(newram > maxram) return res.send("You reached max ram limit!");
       if (newusercoins == 0) {
         await db.delete("coins-" + req.session.userinfo.id);
         await db.set("ram-" + req.session.userinfo.id, newram);
@@ -98,7 +102,7 @@ module.exports.load = async function(app, db) {
 
       let newusercoins = usercoins - cost;
       let newdisk = diskcap + amount;
-
+      if(newdisk > maxdisk) return res.send("You reached max disk limit!");
       if (newusercoins == 0) {
         await db.delete("coins-" + req.session.userinfo.id);
         await db.set("disk-" + req.session.userinfo.id, newdisk);
@@ -160,7 +164,7 @@ module.exports.load = async function(app, db) {
 
       let newusercoins = usercoins - cost;
       let newcpu = cpucap + amount;
-
+      if(newcpu > maxcpu) return res.send("Reached max CPU limit!");
       if (newusercoins == 0) {
         await db.delete("coins-" + req.session.userinfo.id);
         await db.set("cpu-" + req.session.userinfo.id, newcpu);
@@ -222,7 +226,7 @@ module.exports.load = async function(app, db) {
 
       let newusercoins = usercoins - cost;
       let newservers = serverscap + amount;
-
+      if(newservers > maxservers) return res.send("Reached max server limit!");
       if (newusercoins == 0) {
         await db.delete("coins-" + req.session.userinfo.id);
         await db.set("servers-" + req.session.userinfo.id, newservers);
